requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0
feedparser>=6.0.11
python-dotenv>=1.0.1
rich>=13.7.0
requests>=2.32.0
nltk>=3.8.0
streamlit>=1.35.0
plotly>=5.22.0
